Update docker path for DSM v6.0
Use the fully qualified path for the docker binary, to avoid errors
on startup. While the binary should be in the path, it is not
available by the time this script runs. Additionally, the symlink
at /usr/local/bin/docker seems to be overwritten during startup so
it can't be trusted either.

This change closes #3.